requirements.txt
git+https://github.com/Rapptz/discord.py
discord~=1.0.1
requests~=2.21.0
urllib3~=1.24.1
mysql-connector~=2.2.9